k3s no start script

+++ digitalocean/scripts/k3s.sh
@@ -2,17 +2,33 @@
 set -x
 
 export K3S_KUBECONFIG_MODE="644"
-export INSTALL_K3S_EXEC="--flannel-backend=none --cluster-cidr=$K3S_CLUSTER_CIDR --disable-network-policy --disable=traefik"
 
-echo "Using K3S_TOKEN: $K3S_TOKEN"
-echo "Using K3S_CLUSTER_CIDR: $K3S_CLUSTER_CIDR"
-echo "Using K3S_KUBECONFIG_MODE: $K3S_KUBECONFIG_MODE"
-echo "Using INSTALL_K3S_EXEC: $INSTALL_K3S_EXEC"
+echo "==> ENVIRONMENT VARIABLES"
+env
 
+echo "==> Installing K3s Service"
 curl -sfL https://get.k3s.io | sh -
 
-#Prepare kubctl config
-mkdir -p ~/.kube
-cp -n /etc/rancher/k3s/k3s.yaml ~/.kube/config
+if [ -f /etc/systemd/system/k3s.service.env ]; then
+    echo "==> K3s Service ENV"
+    cat /etc/systemd/system/k3s.service.env
+fi
 
-cat ~/.kube/config
+# https://www.bashcookbook.com/bashinfo/source/bash-4.0/examples/scripts/timeout3
+
+if [ -f /etc/systemd/system/k3s-agent.service.env ]; then
+    echo "==> K3s Agent Service ENV"
+    cat /etc/systemd/system/k3s-agent.service.env
+fi
+
+if [ -f /etc/rancher/k3s/k3s.yaml ]; then
+    echo "==> Prepare K3s kubectl"
+    mkdir -p ~/.kube
+    cp -n /etc/rancher/k3s/k3s.yaml ~/.kube/config
+
+    echo "==> K3s kubectl config"
+    cat ~/.kube/config
+
+    echo "==> Copy K3s Node Token to /tmp/server-token"
+    cp /var/lib/rancher/k3s/server/token /tmp/server-token
+fi
